refactor: simplify docker build (#27174)
