+++ .dockerignore
@@ -1,0 +1,8 @@
+*
+!tools/docker/bin
+!dist/
+!node_modules/
+!package.json
+!pnpm-lock.yaml
+!renovate-schema.json
+!license
