+++ package.json
@@ -9,8 +9,8 @@
   },
   "scripts": {
     "build": "run-s clean 'generate:*' 'compile:*' create-json-schema",
-    "build:docker": "node tools/docker.mjs",
-    "build:docs": "run-s 'release:prepare {@}' --",
+    "build:docker": "ts-node tools/docker.ts",
+    "build:docs": "ts-node tools/generate-docs.ts",
     "clean": "rimraf dist tmp",
     "clean-cache": "node tools/clean-cache.mjs",
     "compile:ts": "tsc -p tsconfig.app.json",
@@ -45,8 +45,8 @@
     "pretest": "run-s 'generate:*'",
     "prettier": "prettier --cache --check '**/*.{ts,js,mjs,json,md,yml}'",
     "prettier-fix": "prettier --write --cache '**/*.{ts,js,mjs,json,md,yml}'",
-    "release:prepare": "ts-node tools/generate-docs.ts",
-    "release:publish": "node tools/release.mjs",
+    "release:prepare": "ts-node tools/prepare-release.ts",
+    "release:publish": "ts-node tools/publish-release.ts",
     "start": "ts-node lib/renovate.ts",
     "test": "run-s lint test-schema jest",
     "test-dirty": "git diff --exit-code",
