+++ tools/docker/bin/renovate
@@ -1,0 +1,8 @@
+#!/bin/bash
+
+if [[ -f "/usr/local/etc/env" && -z "${CONTAINERBASE_ENV+x}" ]]; then
+    # shellcheck source=/dev/null
+  . /usr/local/etc/env
+fi
+
+node /usr/local/renovate/dist/renovate.js "$@"
