--- tools/release.mjs
@@ -1,8 +1,0 @@
-import { options } from './utils/options.mjs';
-
-const version = options.release;
-
-console.log(`Publishing version: ${version}`);
-
-// eslint-disable-next-line promise/valid-params,@typescript-eslint/no-floating-promises
-import('./dispatch-release.mjs').catch();
