+++ tools/utils/exec.ts
@@ -1,4 +1,4 @@
-import { spawnSync } from 'node:child_process';
+import { type SpawnSyncReturns, spawnSync } from 'node:child_process';
 
 const maxBuffer = 20 * 1024 * 1024;
 
@@ -7,7 +7,10 @@ const maxBuffer = 20 * 1024 * 1024;
  * @param {string} cmd
  * @param {string[]} args
  */
-export function exec(cmd, args = []) {
+export function exec(
+  cmd: string,
+  args: string[] = [],
+): SpawnSyncReturns<string> {
   // args from shelljs
   return spawnSync(cmd, args, { maxBuffer, encoding: 'utf8' });
 }
